+++ Frontend/src/lib/icons/Search.svelte
@@ -1,0 +1,13 @@
+<script lang="ts">
+  let {className}: {className?: string | undefined} = $props();
+</script>
+
+<svg
+  xmlns="http://www.w3.org/2000/svg"
+  viewBox="0 0 24 24"
+  stroke-width="2"
+  stroke-linecap="round"
+  class={className + ' small-icon fill-none stroke-current'}>
+  <path stroke="none" d="M0 0h24v24H0z" fill="none" /><path d="M10 10m-7 0a7 7 0 1 0 14 0a7 7 0 1 0 -14 0" />
+  <path d="M21 21l-6 -6" />
+</svg>