+++ Frontend/src/lib/icons/index.ts
@@ -2,13 +2,18 @@ export {default as SuccessIcon} from './Success.svelte';
 export {default as AlertIcon} from './Alert.svelte';
 export {default as InfoIcon} from './Info.svelte';
 
-export {default as VerificationIcon} from './Verification.svelte';
-export {default as PasswordIcon} from './Password.svelte';
-export {default as DownloadIcon} from './Download.svelte';
-export {default as RecoveryIcon} from './Recovery.svelte';
-export {default as HappyIcon} from './Happy.svelte';
-export {default as UserIcon} from './User.svelte';
 export {default as CopyIcon} from './Copy.svelte';
+export {default as HappyIcon} from './Happy.svelte';
+export {default as DownloadIcon} from './Download.svelte';
+
+export {default as SearchIcon} from './Search.svelte';
+export {default as FilterIcon} from './Filter.svelte';
+export {default as SortIcon} from './Sort.svelte';
+
+export {default as VerificationIcon} from './Verification.svelte';
+export {default as RecoveryIcon} from './Recovery.svelte';
+export {default as PasswordIcon} from './Password.svelte';
+export {default as UserIcon} from './User.svelte';
 
 export {default as LogoutIcon} from './Logout.svelte';
 export {default as IssuesIcon} from './Issues.svelte';